Add shared CORS utilities
